refactor: add mutation util to reuse mutation parse and verify logic
