+++ src/node/src/lib.rs
@@ -22,6 +22,7 @@ pub mod context;
 pub mod indexer_impl;
 mod json_rpc;
 pub mod json_rpc_impl;
+mod mutation_utils;
 pub mod node_key;
 pub mod node_storage;
 pub mod storage_node_impl;
